requirements.txt
beautifulsoup4==4.12.3
google_api_python_client==2.118.0
google_auth_oauthlib==1.2.0
matplotlib==3.7.3
openai==1.12.0
pandas==1.4.3
pandas==2.0.3
pinecone==0.1.0
protobuf==3.14.0
protobuf==3.13.0
python-dotenv==1.0.1
tiktoken==0.6.0
